refactor: correct path to films
